Type `resolves/rejects` so that LSP recognizes it as awaitable.
Once you use `Expected#resolves` or `#Expected.rejects`, the matcher
functions return a `Promise<void>` and not a `void`. However, the
types do not reflect this and so the language server incorrectly warns
that the 'await' has no effect on the expression.

This maps over the `Expected` type so that if the `resolves` and
`rejects` properties point to a "Promisified" version of `Expected`
where each matcher function function return `Promise<void>` instead of
`void`.