✨Add specific property value match to `toHaveProperty()`
Sometimes you want to test not only that a property exists, but that
it also has a particular value. For example:

```js
let state = getState();
expect(state).toHaveProperty('type', 'ok');
```

This adds an optional `propertyValue` parameter to the
`toHaveProperty()` matcher that will be used to compare against and
fail if it does not match. In the event that this parameter is not
present, the matcher will behave as before.